requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
PyJWT>=2.9.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
email-validator>=2.2.0
